+++ gsl_simulation/views/simulation_projet_views.py
@@ -9,9 +9,9 @@ from django.views.decorators.http import require_POST
 from django.views.generic import DetailView
 
 from gsl.settings import ALLOWED_HOSTS
+from gsl_projet.forms import ProjetForm
 from gsl_projet.services import ProjetService
 from gsl_projet.utils.projet_page import PROJET_MENU
-from gsl_projet.views import ProjetForm
 from gsl_simulation.models import SimulationProjet
 from gsl_simulation.services.simulation_projet_service import (
     SimulationProjetService,
